requirements.txt
# pip 
pygraphviz
pillow

# system
graphviz